# swagger.yml
swagger: '3.0'
info:
  version: 0.0.1
  title: "Foo"
  description: Test Task for Purrweb
schemes: 
- https 
host: test.testo
basePath: "/"
securityDefinitions:
  UserSecurity:
    type: basic
security:
  - UserSecurity: [] 
consumes:
      - "application/json"
produces:
      - "application/json"
paths:
  /users:
    get:
     summary: "Get all users"
     operationId: "getUsers"
     tags:  
      - "users"
     responses:
        "200":
          schema:
            type: "array"
            items: 
              $ref: "#/definitions/User"
          description: "All reistred users"
        "404":
          description: "Users not found"
...
  /users/{id}:
    get:
      summary: "Get User by ID"
      tags:
      - "users"
      operationId: "getUser"
      description: "get user by id"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "User not found"
    delete:
      summary: "Delete user with current ID"
      operationId: "deleteUser"
      tags:
          - "users"
      parameters:    
        - in: "path"
          name: "id"
          type: "integer"
          description: "Change information about user"
          required: true    
      description: "delete user from service"
      responses:
        "200":
          description: "User deleted"
        "400":
          description: "User not found"
...
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
      username:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      columns:
        type: "array"
        items:
          $ref: "#/definitions/Column"
  Column:
    type: "object"
    properties:
      id:
        type: "integer"
      userId:
        type: "integer"
      content:
        type: "string"
      cards: 
        type: "array"
        items: 
          $ref: "#/definitions/Card"
  Card:
    type: "object"
    properties:
     id: 
      type: "integer"
     name: 
      type: "string"
     description:
      type: "string"
     price:
      type: "number"
     comments:
      type: "array"
      items: 
        $ref: "#/definitions/Comment"
  Comment:
    type: "object"
    properties:
      id:
        type: "integer"
      cardID:
        type: "integer"
      content:
        type: "string"